Add files via upload
Nima, who has recently become acquainted with the topic of pointers, decides to solve a problem. But while solving he ran into a problem, he can only work with pointers in the main function. And still not mastered working with them in functions. The problem is to store the prime k even numbers, and their doubles, in two separate arrays with the help of pointers. He after a lot of effort!! He wrote the following code snippet.